Rename AbstractHandler to AbstractErrorHandler
